requirements.txt
datasets==2.18.0
evaluate==0.4.3
transformers==4.39.3
torch==2.1.2
pandas==2.1.4
numpy==1.26.4
nltk==3.9.1
rouge_score==0.1.2
tqdm==4.66.1
wandb==0.16.6
accelerate==0.29.3
huggingface-hub==0.22.2
spacy==3.6.1
pyvi==0.1.1
bert_score==0.3.13
python-dotenv==1.0.0